+++ onefinance/utils.py
@@ -12,6 +12,8 @@ def on_update_purchase_invoice(doc, method):
     res = []
     [res.append(x) for x in po_list if x[0] != None]
     print(res)
+    res = list(set(res))
+    print(res)
     if res:
         doc.set("pre_approved_items", [])
         for i in res:
@@ -132,7 +134,7 @@ def on_submit_journal(doc,method):
 
 def on_update_vendor(doc, method):
     if doc.workflow_state == "Vendor Created":
-        supp = frappe.get_doc({"doctype": "Supplier", "supplier_name": doc.company_name, "supplier_group": "All Supplier Groups", "gst_category": doc.gst_status, "pan": doc.pan_number, "msme_number": doc.vendor_status_on_msme_if_yes_mention_msme_no, "country": doc.country, "website": doc.website})
+        supp = frappe.get_doc({"doctype": "Supplier", "supplier_name": doc.company_name, "supplier_group": "All Supplier Groups", "gst_category": doc.gst_status, "gstin": doc.gst_number, "pan": doc.pan_number, "msme_number": doc.vendor_status_on_msme_if_yes_mention_msme_no, "country": doc.country, "website": doc.website})
         supp.insert(ignore_permissions=True)
 
         add = frappe.new_doc("Address")
@@ -145,6 +147,7 @@ def on_update_vendor(doc, method):
         add.pincode = doc.postal_code
         add.email_id = doc.email_address
         add.phone = doc.landline_number
+        add.gst_category = doc.gst_status
         add.gstin = doc.gst_number
         add.append('links', {
             'link_doctype': 'Supplier',
@@ -167,16 +170,25 @@ def on_update_vendor(doc, method):
         cont.insert()
 
         bank = frappe.new_doc("Bank Account")
-        bank.bank_account_no = doc.bank_account_number
-        bank.account_type = doc.account_type
-        bank.branch_code = doc.ifsc_code
-        bank.bank = doc.bank_name
-        bank.account_name = doc.beneficiary_name
-        bank.party_type = "Supplier"
-        bank.party = doc.company_name
-
-        bank.insert()
-
+        if doc.bank_name != "Other":
+            bank.bank_account_no = doc.bank_account_number
+            bank.branch_code = doc.ifsc_code
+            bank.bank = doc.bank_name
+            bank.account_name = doc.beneficiary_name
+            bank.party_type = "Supplier"
+            bank.party = doc.company_name
+            bank.insert()
+        if doc.bank_name == "Other":
+            new_bank = frappe.new_doc("Bank")
+            new_bank.bank_name = doc.enter_bank_name
+            new_bank.insert()
+            bank.bank_account_no = doc.bank_account_number
+            bank.branch_code = doc.ifsc_code
+            bank.bank = doc.enter_bank_name
+            bank.account_name = doc.beneficiary_name
+            bank.party_type = "Supplier"
+            bank.party = doc.company_name
+            bank.insert()
 
 
         file = frappe.new_doc("File")